requirements.txt
selenium
pytest
openai
python-dotenv
webdriver-manager
beautifulsoup4
